调整server clinet关系
